refactor(compile): extract canonical agentic-pipeline shape into agentic_pipeline.rs
Move every per-job builder, condition builder, variable hoist, dependency
wiring, helper step builder, StandaloneCtx, JobPrefix, BuiltPipelineContext,
build_pipeline_context, and build_canonical_jobs out of standalone_ir.rs
(2,385 lines) into a new src/compile/agentic_pipeline.rs.

standalone_ir.rs shrinks back to its one-screen wrapper, parallel to the
existing job_ir.rs / stage_ir.rs / onees_ir.rs target wrappers. Imports in
those wrappers retargeted at super::agentic_pipeline::build_pipeline_context.

Doc-comment references in extensions/ado_script.rs and extensions/exec_context/
updated to the new module path.

Pure file move — zero behaviour change. All 1832 tests pass, no .yml fixture
drift, clippy clean.

Part of the work for issue #987.
