+++ src/compile/stage_ir.rs
@@ -27,7 +27,7 @@
 //!
 //! Job-id prefixing matches the legacy template (Agent / Detection /
 //! SafeOutputs are prefixed; Setup / Teardown are unprefixed). See
-//! [`crate::compile::standalone_ir::JobPrefix`] for the prefix rule.
+//! [`crate::compile::agentic_pipeline::JobPrefix`] for the prefix rule.
 
 use anyhow::Result;
 use std::path::Path;
@@ -37,7 +37,7 @@ use super::extensions::{CompileContext, Extension};
 use super::ir::ids::StageId;
 use super::ir::stage::{Stage, StageExternalParamsWrap};
 use super::ir::{Pipeline, PipelineBody, PipelineShape, Resources, TemplateParams, Triggers};
-use super::standalone_ir::build_pipeline_context;
+use super::agentic_pipeline::build_pipeline_context;
 use super::types::FrontMatter;
 
 /// Build the typed [`Pipeline`] for the `target: stage` compile
